src/functions/eventhub-to-appinsights/requirements.txt
# Azure Functions
azure-functions==1.18.0
azure-functions-worker==1.0.0

# Azure SDK
azure-identity==1.15.0
azure-eventhub==5.11.4
azure-storage-blob==12.19.0

# Application Insights
azure-monitor-opentelemetry==1.2.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opencensus-ext-azure==1.1.13
applicationinsights==0.11.10

# JSON handling and utilities
requests==2.31.0
python-dateutil==2.8.2

# Configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Logging
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
flake8==6.1.0